refacto test, move test folder
